Minor reworks and changes

+++ Shaders/VertexShader.glsl
@@ -1,7 +1,7 @@
 #version 450 core
 
 layout(location = 0) in vec3 position;
-layout(location = 1) in vec3 instance_position;
+layout(location = 1) in ivec3 instance_position;
 
 uniform mat4 MVP;
